elt-pipeline/requirements.txt
# Python Dependencies for ELT Pipeline

# Core Data Processing
pyspark==3.3.0
pandas==1.5.3
numpy==1.24.0
pyarrow==11.0.0

# Kafka
kafka-python==2.0.2
confluent-kafka==2.1.1

# Database
psycopg2-binary==2.9.5
sqlalchemy==2.0.15

# Airflow
apache-airflow==2.7.1
apache-airflow-providers-apache-spark==4.1.3
apache-airflow-providers-apache-hive==6.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0
requests==2.31.0

# Logging and Monitoring
colorlog==6.7.0

# Testing (optional)
pytest==7.3.1
pytest-cov==4.1.0